refactor(core): R0 split Schema-resolution sum type into its own contract module
workflow.ts carried six orthogonal intents, past the repository's
five-intent ownership limit, because the schemas sum type had grown
inside it. The CliSchemaInfo shape, success array, 1.9 selected-Root
failure envelope, union, and failure discriminator now live in
cli-contracts/schema-resolution.ts with a truthful four-intent header;
workflow.ts returns to five intents and every importer (executor,
opsx-types, opsx-kernel, both contract test files, the barrel index)
points at the focused module. The changed static-provider test gains
its required timestamped Orthogonal intents and v9 Original request
header. Focused: core workflow contract tests and the web static
provider unit suite both pass unchanged in behavior.
