+++ packages/core/src/cli-contracts/index.ts
@@ -1,6 +1,6 @@
 /**
  * Orthogonal intents (created 2026-07-16 Asia/Shanghai):
- * 1. Expose the public OpenSpec CLI 1.6 command-contract surface.
+ * 1. Expose the public OpenSpec CLI 1.6-1.9 command-contract surface.
  * 2. Keep command evidence, common facts, Store facts, and workflow facts discoverable together.
  *
  * Original request (2026-07-15): "为不同命令建立强类型适配器，不实现平行解析规则。"
@@ -8,5 +8,6 @@
 export * from './command-result.js'
 export * from './common.js'
 export * from './executor.js'
+export * from './schema-resolution.js'
 export * from './store.js'
 export * from './workflow.js'
